+++ jgclark.DashboardReact/src/react/customHooks/useInteractiveProcessing.jsx
@@ -30,7 +30,7 @@ function useInteractiveProcessing(
 useEffect(() => {
     if (!reactSettings) return
     if (!reactSettings.interactiveProcessing) return
-    if (reactSettings.interactiveProcessing !== thisSection.name) return
+    if (reactSettings.interactiveProcessing.sectionName !== thisSection.name) return
     logDebug('useInteractiveProcessing', `reactSettings.interactiveProcessing: ${reactSettings?.interactiveProcessing ? 'yes' : 'no'}; items in section ${String(thisSection.name)}: ${items.length} items, itemsCopy (overdue loop items to work on): ${itemsCopy.length} items`)
     if (reactSettings?.interactiveProcessing && items.length > 0 && itemsCopy.length === 0) {
         logDebug('useInteractiveProcessing', `Initializing itemsCopy to: ${items.length} items`)
@@ -44,34 +44,35 @@ useEffect(() => {
   useEffect(() => {
     if (!reactSettings) return
     if (!reactSettings.interactiveProcessing) return
-    if (reactSettings.interactiveProcessing !== thisSection.name) return
+    const { interactiveProcessing } = reactSettings
+    if (interactiveProcessing.sectionName !== thisSection.name) return
     if (itemsCopy.length === 0) return
-    logDebug('useInteractiveProcessing', 'overdue processing, check for items to process', reactSettings)
-    const currentOverdueIndex = reactSettings.currentOverdueIndex || 0
+    logDebug('useInteractiveProcessing', 'overdue processing, check for items to process', reactSettings, interactiveProcessing)
+    const currentIPIndex = interactiveProcessing?.currentIPIndex || 0
     const dialogIsOpen = reactSettings.dialogData?.isOpen || false
 
-    logDebug('useInteractiveProcessing', `currentOverdueIndex: ${currentOverdueIndex}, dialogIsOpen: ${dialogIsOpen} itemsCopy.length: ${itemsCopy.length}`, itemsCopy[currentOverdueIndex])
-    if (!dialogIsOpen && currentOverdueIndex < itemsCopy.length && itemsCopy[currentOverdueIndex]) {
-      logDebug('useInteractiveProcessing', `Opening next overdue item at index: ${currentOverdueIndex}`)
+    logDebug('useInteractiveProcessing', `currentIPIndex: ${currentIPIndex}, dialogIsOpen: ${dialogIsOpen} itemsCopy.length: ${itemsCopy.length}`, itemsCopy[currentIPIndex])
+    if (!dialogIsOpen && currentIPIndex < itemsCopy.length && itemsCopy[currentIPIndex]) {
+      logDebug('useInteractiveProcessing', `Opening next overdue item at index: ${currentIPIndex}`)
       setReactSettings((prev) => ({
         ...prev,
         dialogData: {
           ...prev.dialogData,
           isOpen: true,
           isTask: true,
-          details: { ...prev.dialogData.details, item: itemsCopy[currentOverdueIndex] },
+          details: { ...prev.dialogData.details, item: itemsCopy[currentIPIndex] },
         },
-        currentOverdueIndex: currentOverdueIndex + 1,
+        interactiveProcessing: {...prev.interactiveProcessing,currentIPIndex: currentIPIndex + 1}
       }))
       const actionType = 'showLineInEditorFromFilename'
-      sendActionToPlugin(actionType, { actionType, item: itemsCopy[currentOverdueIndex] }, 'Title clicked in Dialog', true)
-    } else if (!dialogIsOpen && currentOverdueIndex >= itemsCopy.length) {
-        logDebug('useInteractiveProcessing', `Over the limit?  ${String(currentOverdueIndex >= itemsCopy.length)} ${currentOverdueIndex.toString()} >= ${itemsCopy.length.toString()}`)
+      sendActionToPlugin(actionType, { actionType, item: itemsCopy[currentIPIndex] }, 'Title clicked in Dialog', true)
+    } else if (!dialogIsOpen && currentIPIndex >= itemsCopy.length) {
+        logDebug('useInteractiveProcessing', `Over the limit?  ${String(currentIPIndex >= itemsCopy.length)} ${currentIPIndex.toString()} >= ${itemsCopy.length.toString()}`)
         if (itemsCopy?.length > 0) {
             setReactSettings((prev) => ({
                 ...prev,
                 interactiveProcessing: false,
-                currentOverdueIndex: -1,
+                currentIPIndex: -1,
                 dialogData: { isOpen: false, details: null },
             }))
             setItemsCopy([])